# Conflicts:
#	WebHostLib/static/assets/gameInfo/en_Raft.md
#	WebHostLib/static/assets/tutorial/raft/setup_en.md
#	worlds/raft/Options.py
#	worlds/raft/__init__.py
#	worlds/raft/locations.json

+++ worlds/sm/__init__.py
@@ -2,7 +2,7 @@ import logging
 import copy
 import os
 import threading
-from typing import Set, List
+from typing import Set
 
 logger = logging.getLogger("Super Metroid")
 
@@ -11,12 +11,11 @@ from .Items import lookup_name_to_id as items_lookup_name_to_id
 from .Regions import create_regions
 from .Rules import set_rules, add_entrance_rule
 from .Options import sm_options
-from .Rom import get_base_rom_path, ROM_PLAYER_LIMIT
+from .Rom import get_base_rom_path, ROM_PLAYER_LIMIT, SMDeltaPatch
 import Utils
 
 from BaseClasses import Region, Entrance, Location, MultiWorld, Item, RegionType, CollectionState
-from ..AutoWorld import World
-import Patch
+from ..AutoWorld import World, AutoLogicRegister
 
 from logic.smboolmanager import SMBoolManager
 from graph.vanilla.graph_locations import locationsDict
@@ -26,6 +25,32 @@ from rando.Items import ItemManager
 from utils.parameters import *
 from logic.logic import Logic
 from randomizer import VariaRandomizer
+from utils.doorsmanager import DoorsManager
+from rom.rom_patches import RomPatches
+
+
+class SMCollectionState(metaclass=AutoLogicRegister):
+    def init_mixin(self, parent: MultiWorld):
+        
+        # for unit tests where MultiWorld is instantiated before worlds
+        if hasattr(parent, "state"):
+            self.smbm = {player: SMBoolManager(player, parent.state.smbm[player].maxDiff,
+                                    parent.state.smbm[player].onlyBossLeft) for player in
+                                        parent.get_game_players("Super Metroid")}
+            for player, group in parent.groups.items():
+                if (group["game"] == "Super Metroid"):
+                    self.smbm[player] = SMBoolManager(player)
+                    if player not in parent.state.smbm:
+                        parent.state.smbm[player] = SMBoolManager(player)
+        else:
+            self.smbm = {}
+
+    def copy_mixin(self, ret) -> CollectionState:
+        ret.smbm = {player: copy.deepcopy(self.smbm[player]) for player in self.smbm}
+        return ret
+
+    def get_game_players(self, multiword: MultiWorld, game_name: str):
+        return tuple(player for player in multiword.get_all_ids() if multiword.game[player] == game_name)
 
 
 class SMWorld(World):
@@ -44,38 +69,12 @@ class SMWorld(World):
     itemManager: ItemManager
 
     locations = {}
-    hint_blacklist = {'Nothing', 'NoEnergy'}
 
     Logic.factory('vanilla')
 
     def __init__(self, world: MultiWorld, player: int):
         self.rom_name_available_event = threading.Event()
         super().__init__(world, player)
-        
-    def __new__(cls, world, player):
-        
-        # Add necessary objects to CollectionState on initialization
-        orig_init = CollectionState.__init__
-        orig_copy = CollectionState.copy
-
-        def sm_init(self, parent: MultiWorld):
-            if (hasattr(parent, "state")): # for unit tests where MultiWorld is instanciated before worlds
-                self.smbm = {player: SMBoolManager(player, parent.state.smbm[player].maxDiff, parent.state.smbm[player].onlyBossLeft) for player in parent.get_game_players("Super Metroid")}
-            orig_init(self, parent)
-
-
-        def sm_copy(self):
-            ret = orig_copy(self)
-            ret.smbm = {player: copy.deepcopy(self.smbm[player]) for player in self.world.get_game_players("Super Metroid")}
-            return ret
-
-        CollectionState.__init__ = sm_init
-        CollectionState.copy = sm_copy
-
-        if world:
-            world.state.smbm = {}
-
-        return super().__new__(cls)
 
     def generate_early(self):
         Logic.factory('vanilla')
@@ -86,10 +85,12 @@ class SMWorld(World):
         # keeps Nothing items local so no player will ever pickup Nothing
         # doing so reduces contribution of this world to the Multiworld the more Nothing there is though
         self.world.local_items[self.player].value.add('Nothing')
-        self.world.local_items[self.player].value.add('NoEnergy')
+        self.world.local_items[self.player].value.add('No Energy')
 
         if (self.variaRando.args.morphPlacement == "early"):
             self.world.local_items[self.player].value.add('Morph')
+
+        self.remote_items = self.world.remote_items[self.player]
 
         if (len(self.variaRando.randoExec.setup.restrictedLocs) > 0):
             self.world.accessibility[self.player] = self.world.accessibility[self.player].from_text("items")
@@ -165,7 +166,8 @@ class SMWorld(World):
 
     def get_required_client_version(self):
         # changes to client DeathLink handling for 0.2.1
-        return max(super(SMWorld, self).get_required_client_version(), (0, 2, 1))
+        # changes to client Remote Item handling for 0.2.6
+        return max(super(SMWorld, self).get_required_client_version(), (0, 2, 6))
 
     def getWord(self, w):
         return (w & 0x00FF, (w & 0xFF00) >> 8)
@@ -285,6 +287,7 @@ class SMWorld(World):
         openTourianGreyDoors = {0x07C823 + 5: [0x0C], 0x07C831 + 5: [0x0C]}
 
         deathLink = {0x277f04: [self.world.death_link[self.player].value]}
+        remoteItem = {0x277f06: self.getWordArray(0b001 + (0b010 if self.remote_items else 0b000))}
 
         playerNames = {}
         playerNameIDMap = {}
@@ -299,6 +302,7 @@ class SMWorld(World):
                         'offworldSprites': offworldSprites,
                         'openTourianGreyDoors': openTourianGreyDoors,
                         'deathLink': deathLink,
+                        'remoteItem': remoteItem,
                         'PlayerName':  playerNames,
                         'PlayerNameIDMap':  playerNameIDMap}
         romPatcher.applyIPSPatchDict(patchDict)
@@ -308,7 +312,10 @@ class SMWorld(World):
         from Main import __version__
         self.romName = bytearray(f'SM{__version__.replace(".", "")[0:3]}_{self.player}_{self.world.seed:11}\0', 'utf8')[:21]
         self.romName.extend([0] * (21 - len(self.romName)))
+        # clients should read from 0x7FC0, the location of the rom title in the SNES header.
+        # duplicative ROM name at 0x1C4F00 is still written here for now, since people with archipelago pre-0.3.0 client installed will still be depending on this location for connecting to SM
         romPatcher.applyIPSPatch('ROMName', { 'ROMName':  {0x1C4F00 : self.romName, 0x007FC0 : self.romName} })
+
 
         startItemROMAddressBase = 0x2FD8B9
 
@@ -393,23 +400,25 @@ class SMWorld(World):
         romPatcher.writeRandoSettings(self.variaRando.randoExec.randoSettings, itemLocs)
 
     def generate_output(self, output_directory: str):
+        outfilebase = 'AP_' + self.world.seed_name
+        outfilepname = f'_P{self.player}'
+        outfilepname += f"_{self.world.player_name[self.player].replace(' ', '_')}"
+        outputFilename = os.path.join(output_directory, f'{outfilebase}{outfilepname}.sfc')
+
         try:
-            outfilebase = 'AP_' + self.world.seed_name
-            outfilepname = f'_P{self.player}'
-            outfilepname += f"_{self.world.player_name[self.player].replace(' ', '_')}" \
-
-            outputFilename = os.path.join(output_directory, f'{outfilebase}{outfilepname}.sfc')
             self.variaRando.PatchRom(outputFilename, self.APPatchRom)
-
             self.write_crc(outputFilename)
-
-            Patch.create_patch_file(outputFilename, player=self.player, player_name=self.world.player_name[self.player], game=Patch.GAME_SM)
-            os.unlink(outputFilename)
             self.rom_name = self.romName
         except:
             raise
+        else:
+            patch = SMDeltaPatch(os.path.splitext(outputFilename)[0]+SMDeltaPatch.patch_file_ending, player=self.player,
+                                 player_name=self.world.player_name[self.player], patched_path=outputFilename)
+            patch.write()
         finally:
-            self.rom_name_available_event.set() # make sure threading continues and errors are collected
+            if os.path.exists(outputFilename):
+                os.unlink(outputFilename)
+            self.rom_name_available_event.set()  # make sure threading continues and errors are collected
 
     def checksum_mirror_sum(self, start, length, mask = 0x800000):
         while (not(length & mask) and mask):
@@ -425,8 +434,6 @@ class SMWorld(World):
             while (next_length < mask):
                 next_length += next_length
                 part2 += part2
-
-            length = mask + mask
 
         return (part1 + part2) & 0xFFFF
 
@@ -450,21 +457,45 @@ class SMWorld(World):
         # we skip in case of error, so that the original error in the output thread is the one that gets raised
         if rom_name:
             new_name = base64.b64encode(bytes(self.rom_name)).decode()
-            payload = multidata["connect_names"][self.world.player_name[self.player]]
-            multidata["connect_names"][new_name] = payload
-            del (multidata["connect_names"][self.world.player_name[self.player]])
-
+            multidata["connect_names"][new_name] = multidata["connect_names"][self.world.player_name[self.player]]
 
     def fill_slot_data(self): 
         slot_data = {}
+        if not self.world.is_race:
+            for option_name in self.options:
+                option = getattr(self.world, option_name)[self.player]
+                slot_data[option_name] = option.value
+
+            slot_data["Preset"] = { "Knows": {},
+                                    "Settings": {"hardRooms": Settings.SettingsDict[self.player].hardRooms,
+                                                 "bossesDifficulty": Settings.SettingsDict[self.player].bossesDifficulty,
+                                                 "hellRuns": Settings.SettingsDict[self.player].hellRuns},
+                                    "Controller": Controller.ControllerDict[self.player].__dict__}
+
+            for knows in Knows.__dict__:
+                if isKnows(knows):
+                    slot_data["Preset"]["Knows"][knows] = [ getattr(Knows.knowsDict[self.player], knows).bool, 
+                                                            getattr(Knows.knowsDict[self.player], knows).difficulty]
+
+            slot_data["InterAreaTransitions"] = {}
+            for src, dest in self.variaRando.randoExec.areaGraph.InterAreaTransitions:
+                slot_data["InterAreaTransitions"][src.Name] = dest.Name
+                
+            slot_data["Doors"] = {}
+            for door in DoorsManager.doorsDict[self.player].values():
+                slot_data["Doors"][door.name] = door.getColor()
+
+            slot_data["RomPatches"] = RomPatches.ActivePatches[self.player]
+                
         return slot_data
 
     def collect(self, state: CollectionState, item: Item) -> bool:
-        state.smbm[item.player].addItem(item.type)
-        if item.advancement:
-            state.prog_items[item.name, item.player] += 1
-            return True  # indicate that a logical state change has occured
-        return False
+        state.smbm[self.player].addItem(item.type)
+        return super(SMWorld, self).collect(state, item)
+
+    def remove(self, state: CollectionState, item: Item) -> bool:
+        state.smbm[self.player].removeItem(item.type)
+        return super(SMWorld, self).remove(state, item)
 
     def create_item(self, name: str) -> Item:
         item = next(x for x in ItemManager.Items.values() if x.Name == name)
@@ -499,6 +530,8 @@ class SMWorld(World):
         for item in self.world.itempool:
             if item.player == self.player and item.advancement:
                 progitempool.append(item)
+            if item.location.player == self.player and item.game == "Super Metroid" and item.type == "Nothing":
+                item.location.address = None
 
         for item in progitempool:
             new_state.collect(item, True)
@@ -509,9 +542,11 @@ class SMWorld(World):
                 self.world.state.smbm[self.player].onlyBossLeft = True
                 break
 
+
 def create_locations(self, player: int):
     for name, id in locations_lookup_name_to_id.items():
         self.locations[name] = SMLocation(player, name, id)
+
 
 def create_region(self, world: MultiWorld, player: int, name: str, locations=None, exits=None):
     ret = Region(name, RegionType.LightWorld, name, player)
